requirements.txt
opencv-python==4.6.0.66
Pillow==9.2.0
numpy==1.23.4